index changes
Not sure but we are iterating through pairs of galaxies in a triangular manner, starting from index + 1 in the inner loop. This avoids redundant calculations, but it means we are only considering one pair (i, j) but not the pair (j, i). This may not be suitable if the order of galaxies matters in your calculation. If wewant to include both (i, j) and (j, i) pairs, we wikk need to adjust the loop to start from 0 in the inner loop and avoid counting the pair twice.